Refactoring Cassandra workflow persistence manager for NoSQL support-Part 3
