+++ common/persistence/nosql/factory.go
@@ -27,8 +27,6 @@ import (
 	"github.com/uber/cadence/common/log"
 	p "github.com/uber/cadence/common/persistence"
 
-	// NOTE: this package will be refactored and removed soon
-	cassandraOld "github.com/uber/cadence/common/persistence/cassandra"
 	"github.com/uber/cadence/common/persistence/nosql/nosqlplugin/cassandra"
 	"github.com/uber/cadence/common/persistence/nosql/nosqlplugin/cassandra/gocql"
 )
@@ -152,7 +150,7 @@ func (f *executionStoreFactory) close() {
 
 // new implements ExecutionStoreFactory interface
 func (f *executionStoreFactory) new(shardID int) (p.ExecutionStore, error) {
-	pmgr, err := cassandraOld.NewWorkflowExecutionPersistence(shardID, f.client, f.session, f.logger)
+	pmgr, err := NewExecutionStore(shardID, f.client, f.session, f.logger)
 	if err != nil {
 		return nil, err
 	}
